requirements.txt
textual
requests
python-dotenv